Assignment 3: Update
Unfortunately not a very detailed message.

I have basically gotten to the point of having
an index.html page (or a servlet displaying the same information)
passing the user over to the LogsServlet page which will responded to
data entered into the form and display it by updating the HTML page
(basically re-rendering it) with the data currently being stored in an
array full of TextLog Objects.

and well thats it. two days worth of effort.