Assignment 3 Update:
No big commit message this time. (6-7hr) for 3.5 days to get this
assignment where it was has left me little time.

We are Currently almost 100% feature complete, all thats left is to have
the response message delivered to the user and displayed on the HTML.

Should have done more commits...